bugfixes + added path

+++ model/post.js
@@ -6,6 +6,7 @@ const responses = require('../responses')
 
 const pollModel = require('./poll')
 const imageModel = require('./image')
+const pathModel = require('./path')
 
 const postModel = module.exports = {
     doesPostExists: postId => new Promise(async (resolve, reject) => {
@@ -24,6 +25,7 @@ const postModel = module.exports = {
             const post = await postPromise
             const poll = post.pollId ? await pollModel.getOptionsForPoll(post.pollId) : null
             const image = post.imageId ? await imageModel.getImageById(post.imageId) : null
+            const path = post.pathId ? await pathModel.getPathById(post.pathId) : null
             const likes = await postModel.getLikesForPost(post.id)
             const comments = await postModel.getCommentsForPost(post.id)
 
@@ -33,6 +35,7 @@ const postModel = module.exports = {
                 creationDate: post.creationDate,
                 poll: poll,
                 image: image,
+                path: path,
                 likes: likes,
                 comments: comments
             }
@@ -52,15 +55,18 @@ const postModel = module.exports = {
             const post = posts[0]
             const poll = post.pollId ? await pollModel.getOptionsForPoll(post.pollId) : null
             const image = post.imageId ? await imageModel.getImageById(post.imageId) : null
+            const path = post.pathId ? await pathModel.getPathById(post.pathId) : null
             const likes = await postModel.getLikesForPost(postId)
             const comments = await postModel.getCommentsForPost(postId)
 
             const fullPost = {
                 id: post.id,
+                groupId: post.groupId,
                 content: post.content,
                 creationDate: post.creationDate,
                 poll: poll,
                 image: image,
+                path: path,
                 likes: likes,
                 comments: comments
             }
@@ -71,8 +77,10 @@ const postModel = module.exports = {
         }
     }),
 
-    userHasPerMissionToViewPost: (userId, postId) => new Promise(async (resolve, reject) => {
+    userHasPerMissionToViewPost: (postId, userId) => new Promise(async (resolve, reject) => {
         try {
+
+            console.log(`userId: ${userId}, postId: ${postId}`)
 
             const users = await query(`SELECT usersInGroup.id
                          FROM usersInGroup
@@ -150,14 +158,26 @@ const postModel = module.exports = {
 
                     }
 
-                    if (typeof post.image !== "undefined" && post.image) {
+                    if (typeof post.image !== "undefined" && post.image !== null) {
                         await queryWithConnection(connection, `INSERT INTO image (path) VALUES (?)`, [post.image])
                         imageId = (await queryWithConnection(connection, `SELECT LAST_INSERT_ID() AS id`))[0]['id']
                     }
 
-                    await queryWithConnection(connection, `INSERT INTO post (groupId, content, imageId, pollId, creator)
-                                                           VALUES (?, ?, ?, ?, ?)`,
-                        [groupId, post.content, imageId, pollId, userId])
+                    let pathId = null
+                    if (post.hasOwnProperty('path') && post.path !== null && post.path.length > 0) {
+                        await queryWithConnection(connection, `INSERT INTO path () VALUES ()`, [])
+                        pathId = (await queryWithConnection(connection, `SELECT LAST_INSERT_ID() as id`))[0]['id']
+
+                        const insertStr = post.path.map(pathElement => `(?, ?, ?)`).join(',')
+                        const insertValues = flatMap(post.path, pathElement => [pathId, pathElement.x, pathElement.y])
+
+                        await queryWithConnection(connection, `INSERT INTO pathElements (pathId, x, y)
+                                                               VALUES ${insertStr}`, insertValues)
+                    }
+
+                    await queryWithConnection(connection, `INSERT INTO post (groupId, content, imageId, pollId, pathId, creator)
+                                                           VALUES (?, ?, ?, ?, ?, ?)`,
+                        [groupId, post.content, imageId, pollId, pathId, userId])
 
                     const postId = (await queryWithConnection(connection, `SELECT LAST_INSERT_ID() AS id`))[0]['id']
 
@@ -185,9 +205,15 @@ const postModel = module.exports = {
 
             const pollIds = await query(`SELECT pollId FROM post WHERE id = ?`, [postId])
 
+            console.log(`pollids: ${JSON.stringify(pollIds)}`)
+
             if (pollIds.length === 0) return reject(responses.notFound())
 
-            const pollId = pollIds[0]
+            const pollId = pollIds[0]['pollId']
+
+            if (pollId === null) return reject(responses.notFound())
+
+            console.log(`pollId: ${JSON.stringify(pollId)}`)
 
             await query(`DELETE FROM vote WHERE pollId = ? AND voter = ?`, [pollId, userId])
 
@@ -206,6 +232,8 @@ const postModel = module.exports = {
             const posts = await query(`SELECT id FROM post WHERE creator = ? AND id = ?`,
                 [userId, postId])
 
+            console.log(`user: ${userId}, post: ${postId} result: ${JSON.stringify(posts)}`)
+
             resolve(posts.length > 0)
 
         } catch (error) {
@@ -217,46 +245,147 @@ const postModel = module.exports = {
         transaction(async connection => {
             try {
 
-                let pollId = null
-                let imageId = null
+                console.log(`oldPost: ${JSON.stringify(oldPost)}`)
 
-                if (typeof post.poll !== "undefined" && post.poll) {
-                    await queryWithConnection(connection, `INSERT INTO poll (question)
-                                                               VALUES (?)`, [post.poll.question])
-                    pollId = (await queryWithConnection(connection, `SELECT LAST_INSERT_ID() AS id`))[0]['id']
+                if (post.poll !== oldPost.poll) {
+                    if (!post.hasOwnProperty('poll') || post.poll === null) {
+                        const pollId = (await queryWithConnection(connection,
+                            `SELECT pollId FROM post WHERE id = ?`, [postId]))[0]['pollId']
+                        await queryWithConnection(connection, `UPDATE post SET pollId = NULL WHERE id = ?`, [postId])
+                        await queryWithConnection(connection, `DELETE FROM pollOption WHERE pollId = ?`, [pollId])
+                        await queryWithConnection(connection, `DELETE FROM vote WHERE pollId = ?`, [pollId])
+                        await queryWithConnection(connection, `DELETE FROM poll WHERE id = ?`, [pollId])
+                    } else if (typeof oldPost.poll !== "undefined" && oldPost.poll !== null) {
+                        await queryWithConnection(connection, `UPDATE poll
+                                                               SET question = ?
+                                                               WHERE id = ?`,
+                            [post.poll.question, post.poll.id])
 
-                    const answersString = post.poll.answers.map(() => '(?, ?)').join(',')
-                    const params = flatMap(post.poll.answers, answer => [pollId, answer])
+                        console.log(`oldPost: ${JSON.stringify(oldPost)}`)
 
-                    await queryWithConnection(connection, `INSERT INTO pollOption (pollId, answer)
+
+                        await queryWithConnection(connection, `DELETE FROM vote WHERE pollId = ?`, [oldPost.poll.id])
+                        await queryWithConnection(connection, `DELETE FROM pollOption WHERE pollId = ?`, [oldPost.poll.id])
+                        await queryWithConnection(connection, `DELETE FROM poll WHERE id = ?`, [oldPost.poll.id])
+
+                        await queryWithConnection(connection, `INSERT INTO poll (question) VALUES (?)`,
+                            [post.poll.question])
+                        const pollId = (await queryWithConnection(connection, `SELECT LAST_INSERT_ID() as id`))[0]['id']
+
+                        const answersString = post.poll.answers.map(() => '(?, ?)').join(',')
+                        const params = flatMap(post.poll.answers, answer => [pollId, answer])
+
+                        await queryWithConnection(connection, `INSERT INTO pollOption (pollId, answer)
                                                                VALUES ${answersString}`, params)
 
+                        await queryWithConnection(connection, `UPDATE post SET pollId = ? WHERE id = ?`,
+                            [pollId, postId])
+
+                    } else {
+
+                        await queryWithConnection(connection, `INSERT INTO poll (question) VALUES (?)`,
+                            [post.poll.question])
+                        const pollId = (await queryWithConnection(connection, `SELECT LAST_INSERT_ID() as id`))[0]['id']
+
+                        const answersString = post.poll.answers.map(() => '(?, ?)').join(',')
+                        const params = flatMap(post.poll.answers, answer => [pollId, answer])
+
+                        await queryWithConnection(connection, `INSERT INTO pollOption (pollId, answer)
+                                                               VALUES ${answersString}`, params)
+
+                        await queryWithConnection(connection, `UPDATE post SET pollId = ? WHERE id = ?`,
+                            [pollId, postId])
+
+                    }
                 }
 
-                if (typeof oldPost.poll !== "undefined" && oldPost.poll) {
-                    await query(`UPDATE vote SET pollId = ? WHERE pollId = ?`,
-                        [oldPost.poll.id, pollId])
+                console.log(`oldPost:\n${JSON.stringify(oldPost, null, 2)}\n post:\n${JSON.stringify(post, null, 2)}`)
 
-                    await query(`DELETE FROM poll WHERE id = ?`, [oldPost.poll.id])
+                if (oldPost.image !== post.image) {
+                    if (!post.hasOwnProperty('image') || post.image === null) {
+
+                        const imageId = (await queryWithConnection(connection,
+                                `SELECT imageId FROM post WHERE id = ?`, [postId]))[0]['imageId']
+                        await queryWithConnection(connection, `UPDATE post SET imageId = NULL WHERE id = ?`, [postId])
+                        await queryWithConnection(connection, `DELETE FROM image WHERE id = ?`, [imageId])
+
+                    } else if (typeof oldPost.image !== "undefined" && oldPost.image !== null) {
+
+                        await queryWithConnection(connection, `UPDATE image SET path = ?
+                                                               WHERE id = (SELECT imageId FROM post WHERE id = ?)`,
+                            [post.image, postId])
+
+                    } else {
+                        await queryWithConnection(connection, `INSERT INTO image (path) VALUES (?)`, [post.image])
+                        const imageId = (await queryWithConnection(connection, `SELECT LAST_INSERT_ID() as id`))[0]['id']
+                        await queryWithConnection(connection, `UPDATE post SET imageId = ? WHERE id = ?`,
+                            [imageId, postId])
+                    }
                 }
 
-                if (typeof post.image !== "undefined" && post.image) {
-                    await queryWithConnection(connection, `INSERT INTO image (path) VALUES (?)`, [post.image])
-                    imageId = (await queryWithConnection(connection, `SELECT LAST_INSERT_ID() AS id`))[0]['id']
+
+                if (post.path !== oldPost.path) {
+                    if (!post.hasOwnProperty('path') || post.path === null || post.path.length === 0) {
+                        const pathId = (await queryWithConnection(connection,
+                            `SELECT pathId FROM post WHERE id = ?`, [postId]))[0]['pathId']
+                        await queryWithConnection(connection, `UPDATE post SET pathId = NULL WHERE id = ?`, [postId])
+                        await queryWithConnection(connection, `DELETE FROM pathElements WHERE pathId = ?`, [pathId])
+                        await queryWithConnection(connection, `DELETE FROM path WHERE id = ?`, [pathId])
+                    } else if (typeof oldPost.path !== "undefined" && oldPost.path !== null && oldPost.path.length > 0) {
+
+                        const oldPathId = (await queryWithConnection(connection,
+                            `SELECT pathId FROM post WHERE id = ?`, [postId]))[0]['pathId']
+
+                        await queryWithConnection(connection, `DELETE FROM pathElements WHERE pathId = ?`, [oldPathId])
+                        await queryWithConnection(connection, `DELETE FROM path WHERE id = ?`, [oldPathId])
+
+                        await queryWithConnection(connection, `INSERT INTO path () VALUES ()`, [])
+                        const pathId = (await queryWithConnection(connection, `SELECT LAST_INSERT_ID() as id`))[0]['id']
+
+                        const answersString = post.path.map(() => '(?, ?, ?)').join(',')
+                        const params = flatMap(post.path, path => [pathId, path.x, path.y])
+
+                        await queryWithConnection(connection, `INSERT INTO pathElements (pathId, x, y)
+                                                               VALUES ${answersString}`, params)
+
+                        await queryWithConnection(connection, `UPDATE post SET pathId = ? WHERE id = ?`,
+                            [pathId, postId])
+
+                    } else {
+
+                        await queryWithConnection(connection, `INSERT INTO path () VALUES ()`, [])
+                        const pathId = (await queryWithConnection(connection, `SELECT LAST_INSERT_ID() as id`))[0]['id']
+
+                        const answersString = post.path.map(() => '(?, ?, ?)').join(',')
+                        const params = flatMap(post.path, path => [pathId, path.x, path.y])
+
+                        await queryWithConnection(connection, `INSERT INTO pathElements (pollId, answer)
+                                                               VALUES ${answersString}`, params)
+
+                        await queryWithConnection(connection, `UPDATE post SET pathId = ? WHERE id = ?`,
+                            [pathId, postId])
+
+                    }
                 }
 
-                await queryWithConnection(connection, `INSERT INTO post (groupId, content, imageId, pollId, creator)
-                                                           VALUES (?, ?, ?, ?, ?)`,
-                    [oldPost.groupId, post.content, imageId, pollId, userId])
 
-                const postId = (await queryWithConnection(connection, `SELECT LAST_INSERT_ID() AS id`))[0]['id']
+                let pathId = null
+                if (post.hasOwnProperty('path') && post.path !== null && post.path.length > 0) {
+                    await queryWithConnection(connection, `INSERT INTO path () VALUES ()`, [])
+                    pathId = (await queryWithConnection(connection, `SELECT LAST_INSERT_ID() as id`))[0]['id']
 
-                await queryWithConnection(connection, `DELETE FROM post WHERE id = ?`,
-                    [oldPost.id])
+                    const insertStr = path.map(pathElement => `(?, ?, ?)`)
+                    const insertValues = flatMap(post.path, pathElement => [pathId, pathElement.x, pathElement.y])
+
+                    await queryWithConnection(connection, `INSERT INTO pathElements (pathId, x, y)
+                                                               VALUES ${insertStr}`, insertValues)
+                }
+
+                await queryWithConnection(connection, `UPDATE post SET content = ? WHERE id = ?`, [post.content, postId])
 
                 connection.commit()
 
-                resolve(postId)
+                resolve()
 
             } catch (error) {
                 connection.rollback()
